demo-app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
# python:3.12-slim no longer ships setuptools, but opentelemetry-instrumentation
# still imports `pkg_resources` from it. setuptools 81+ removed `pkg_resources`,
# so we have to pin below that until OTel switches to importlib.metadata.
setuptools>=69,<81
